Unify `sha1`
