+++ src/crypto/mod.rs
@@ -49,29 +49,10 @@ pub fn sha1_hmac(key: &[u8], payloads: &[&[u8]]) -> [u8; 20] {
     hmac.finalize().into_bytes().into()
 }
 
-/// If openssl is enabled and sha1 is not, it uses `openssl` crate.
-#[cfg(all(feature = "openssl", not(feature = "sha1")))]
-pub fn sha1_hmac(key: &[u8], payloads: &[&[u8]]) -> [u8; 20] {
-    use openssl::hash::MessageDigest;
-    use openssl::pkey::PKey;
-    use openssl::sign::Signer;
-
-    let key = PKey::hmac(key).expect("valid hmac key");
-    let mut signer = Signer::new(MessageDigest::sha1(), &key).expect("valid signer");
-
-    for payload in payloads {
-        signer.update(payload).expect("signer update");
-    }
-
-    let mut hash = [0u8; 20];
-    signer.sign(&mut hash).expect("sign to array");
-    hash
-}
-
-/// If wincrypto is enabled and sha1 is not, it uses `wincrypto` crate.
+/// if sha1 is not enabled, delegate to the crypto impl.
 #[cfg(all(feature = "wincrypto", not(feature = "sha1")))]
 pub fn sha1_hmac(key: &[u8], payloads: &[&[u8]]) -> [u8; 20] {
-    wincrypto::sha1_hmac(key, payloads)
+    _impl::sha1_hmac(key, payloads)
 }
 
 /// Errors that can arise in DTLS.
